requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
albumentations=1.3.0=pyhd8ed1ab_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=pyh71513ae_1
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blosc=1.21.1=hcbbf2c4_0
bottleneck=1.3.5=py39h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py39h2bbff1b_1003
bzip2=1.0.8=h8ffe710_4
ca-certificates=2022.9.24=h5b45459_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py39h2bbff1b_0
cfitsio=3.470=h2bbff1b_7
charls=2.2.0=h6c2663c_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.2.0=pyhd8ed1ab_0
colorama=0.4.5=py39haa95532_0
cryptography=37.0.1=py39h21b164f_0
cudatoolkit=11.6.0=hc0ea762_10
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py39hb82d6ee_0
dask-core=2022.10.0=pyhd8ed1ab_0
debugpy=1.5.1=py39hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
eigen=3.4.0=h2d74725_0
entrypoints=0.4=py39haa95532_0
executing=0.8.3=pyhd3eb1b0_0
fftw=3.3.9=h2bbff1b_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
fsspec=2022.8.2=pyhd8ed1ab_0
giflib=5.2.1=h8d14728_2
glib=2.69.1=h5dc1a3c_1
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.10.6=nompi_he0bbb20_101
icc_rt=2022.1.0=h6049295_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
imagecodecs=2021.8.26=py39hc0a7faf_1
imageio=2.22.0=pyhfa7a67d_0
importlib-metadata=4.11.4=py39hcbf5309_0
importlib_resources=5.9.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.2=py39haa95532_0
ipython=8.4.0=py39haa95532_0
ipywidgets=8.0.2=pyhd8ed1ab_1
jedi=0.18.1=py39haa95532_1
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
jsonschema=4.16.0=pyhd8ed1ab_0
jupyter_client=7.3.5=py39haa95532_0
jupyter_core=4.10.0=py39haa95532_0
jupyterlab_widgets=3.0.3=pyhd8ed1ab_0
kiwisolver=1.4.2=py39hd77b12b_0
lcms2=2.12=h2a16943_0
lerc=3.0=hd77b12b_0
libaec=1.0.6=h39d44d4_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_4
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libtiff=4.4.0=h8a3f274_0
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
libzopfli=1.0.3=h0e60522_0
locket=1.0.0=pyhd8ed1ab_0
lz4-c=1.9.3=h2bbff1b_1
matplotlib=3.5.2=py39haa95532_0
matplotlib-base=3.5.2=py39hd77b12b_0
matplotlib-inline=0.1.6=py39haa95532_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
munkres=1.1.4=py_0
nbformat=5.6.1=pyhd8ed1ab_0
nest-asyncio=1.5.5=py39haa95532_0
networkx=2.8.7=pyhd8ed1ab_0
numexpr=2.8.3=py39hb80d3ca_0
numpy=1.21.5=py39h7a0a035_3
numpy-base=1.21.5=py39hca35cd5_3
opencv=4.6.0=py39h104de81_2
opencv-python=4.6.0.66=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1q=h8ffe710_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.4=py39hd77b12b_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.3.0=pyhd8ed1ab_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py39hdc2b20a_1
pip=22.1.2=py39haa95532_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
ply=3.11=py39haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py39haa95532_0
pyqt=5.15.7=py39hd77b12b_0
pyqt5-sip=12.11.0=py39hd77b12b_0
pyrsistent=0.18.1=py39hb82d6ee_1
pysocks=1.7.1=py39haa95532_0
python=3.9.13=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytorch=1.12.1=py3.9_cuda11.6_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py39haa95532_0
pywavelets=1.3.0=py39h5d4886f_1
pywin32=302=py39h2bbff1b_2
pyyaml=6.0=py39hb82d6ee_4
pyzmq=23.2.0=py39hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h3ad3cdb_4
qudida=0.0.4=pyhd8ed1ab_0
requests=2.28.1=py39haa95532_0
scikit-image=0.19.3=py39h2e25243_1
scikit-learn=1.1.1=py39hd77b12b_0
scipy=1.7.3=py39h7a0a035_2
setuptools=63.4.1=py39haa95532_0
sip=6.6.2=py39hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h82413e6_1
sqlite=3.39.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.4.8=pyhd8ed1ab_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=pyhd8ed1ab_0
torchaudio=0.12.1=py39_cu116
torchvision=0.13.1=py39_cu116
tornado=6.2=py39h2bbff1b_0
tqdm=4.64.1=py39haa95532_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py39haa95532_0
typing_extensions=4.3.0=py39haa95532_0
tzdata=2022c=h04d1e81_0
urllib3=1.26.11=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.3=pyhd8ed1ab_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
xz=5.2.6=h8cc25b3_0
yaml=0.2.5=h8ffe710_2
zeromq=4.3.4=hd77b12b_0
zfp=0.5.5=h0e60522_8
zipp=3.8.1=pyhd8ed1ab_0
zlib=1.2.12=h8cc25b3_3
zstd=1.5.2=h19a0ad4_0
